pyproject.toml
[project]
name = "webscrap-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
    "tinydb>=4.8.2",
]
